+++ packages/src/@terra-money/wallet-provider/impl-react/StaticWalletProvider.tsx
