+++ packages/src/@terra-money/wallet-provider/index.ts
@@ -1,6 +1,6 @@
 export * from './getChainOptions';
 export * from './controller';
-export * from './react';
+export * from './impl-react';
 
 export * from '@terra-money/use-wallet';
 export type { ReadonlyWalletSession } from './modules/readonly-wallet';
